# Use test cache for test result images too.