@tacaswell 2019-12-13
If we have a few tests that are not deterministic, will this cause the cache to increase without bound? 

I think that we cache these folders on at least travis so we would want to make sure we don't start spending more time moving around invalid images than we save by not computing them.