@anntzer 2019-12-13
Fair enough, added a simple cache eviction mechanism.