@jklymak 2020-09-24
I don't fully follow this, but it *seems* useful.  But needs a rebase, and probably a review from @tacaswell or @QuLogic